requirements.txt
blinker==1.8.1
click==8.1.7
Faker==25.0.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.4.0
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==3.0.2
